Add `better_n` module
This rebinds `n` (which is a rather convenient binding), so that it used
for multiple different movement commands, in the same vein `.` repeats
action (?) commands.

For example, if we jump to the next hunk, using `<leader>hn`, we can
repeat that command using `n`, allowing for far easier "scrolling" using
that motion without coming up with a bind that is easier to repeat.

This addresses one of the problems that I have with Vim, which is that
almost every single binding is used by default, for (often) very niche
actions. I want to be able to reuse convenient bindings for similar
things, reducing both mental overhead as well as opening up more
bindings, allowing Vim to be more ergonomic.